Finished section 4

+++ lab1_tools.py
@@ -52,16 +52,16 @@ def trfbank(fs, nfft, lowfreq=133.33, linsc=200/3., logsc=1.0711703, nlinfilt=13
     """Compute triangular filterbank for MFCC computation.
 
     Inputs:
-    fs:         sampling frequency (rate)
-    nfft:       length of the fft
-    lowfreq:    frequency of the lowest filter
-    linsc:      scale for the linear filters
-    logsc:      scale for the logaritmic filters
-    nlinfilt:   number of linear filters
-    nlogfilt:   number of log filters
+        fs:         sampling frequency (rate)
+        nfft:       length of the fft
+        lowfreq:    frequency of the lowest filter
+        linsc:      scale for the linear filters
+        logsc:      scale for the logaritmic filters
+        nlinfilt:   number of linear filters
+        nlogfilt:   number of log filters
 
     Outputs:
-    res:  array with shape [N, nfft], with filter amplitudes for each column.
+        res:  array with shape [N, nfft], with filter amplitudes for each column.
             (N=nlinfilt+nlogfilt)
     From scikits.talkbox"""
     # Total number of filters
@@ -90,10 +90,10 @@ def trfbank(fs, nfft, lowfreq=133.33, linsc=200/3., logsc=1.0711703, nlinfilt=13
         hi = freqs[i+2]
 
         lid = np.arange(np.floor(low * nfft / fs) + 1,
-                        np.floor(cen * nfft / fs) + 1, dtype=np.int)
+                        np.floor(cen * nfft / fs) + 1, dtype=np.int16)
         lslope = heights[i] / (cen - low)
         rid = np.arange(np.floor(cen * nfft / fs) + 1,
-                        np.floor(hi * nfft / fs) + 1, dtype=np.int)
+                        np.floor(hi * nfft / fs) + 1, dtype=np.int16)
         rslope = heights[i] / (hi - cen)
         fbank[i][lid] = lslope * (nfreqs[lid] - low)
         fbank[i][rid] = rslope * (hi - nfreqs[rid])
